Bug 1896745: Avoid unnecessary refcount traffic when appending DOMRect & PaintRequest to array-of-RefPtrs. r=mccr8
For these objects, we have an `Append(Foo*)` method which is just a
wrapper-function that appends to a `nsTArray<RefPtr<Foo>>`.  As it happens,
every callsite is passing in a local `RefPtr` variable that's no longer needed
after the call to `Append`.

It's slightly more performant to use move-semantics to directly transfer
ownership of the existing reference from the local variable to the array,
rather than having the reference count take one step forward and then one step
back, as we were doing before this patch.

Differential Revision: https://phabricator.services.mozilla.com/D210364

UltraBlame original commit: 6b4cc2f090dc7e22fceab543a9d730aedf733ff2