Bug 1896652 - [devtools] Fix layout glitch in storage table and variable view. r=devtools-reviewers,ochameau.
The plain class that was removed in Bug 1896330 was set
on a few elements in the variable view, the table and
the chart widgets, as well as in the inspector breadcrumb.
This patch removes the plain class and adjust styling
where it's needed.

Differential Revision: https://phabricator.services.mozilla.com/D210316

UltraBlame original commit: 700a89ff1b0dc8d0bc6822c8e3ae0ce30785ce12