Bug 1896738 - Avoid unnecessary clears when using PutImageData on a new canvas. r=aosmond
This fixes two issues. The first is that the dirty rect passed to EnsureTarget was being
transformed, whereas PutImageData is supposed to bypass transforms. The second is that
PutImageData actually initializes the alpha channel properly, such that Skia DTs do not
need the alpha channel of pixels that are modified by PutImageData to be cleared first,
so that PutImageData can effectively function as the clear itself.

Differential Revision: https://phabricator.services.mozilla.com/D210408

UltraBlame original commit: 8a97df11ef6fe5685d068f4e80ba94864c5abf0b