Bug 1875502 - Pre: Add separate XPCOM interface for internal use of the update manager r=nalexander,application-update-reviewers
Looking forward, a goal of this patch stack is to have external interfaces automatically initialize the update system before doing anything that requires that. However, during that initialization, we need to make sure that we don't accidentally cause recursion by invoking an external interface that also does initialization. To do this, we will require special, privileged access to this functionality that doesn't do the initialization.

An appealing way of implementing this is to have `UpdateManager` implement both `nsIUpdateManager` and `nsIUpdateManagerInternal`. However, by experimentation, this seems to allow us to use the internal functions directly on the update manager object, even without ever QIing it to `nsIUpdateManagerInternal`. This is undesirable both because it allows consumers to use internal interfaces sort of implicitly and because it means that if we move functionality from the public interface to the private one, existing consumers of that interface could more easily be missed since the method is still available on the update manager object that it's already using.

This commit doesn't actually add any attributes or methods to the new interface. That will be done later in the patch stack.

Differential Revision: https://phabricator.services.mozilla.com/D209106

UltraBlame original commit: 1950b330d253d094dcf706a489ca13484345be60