Bug 1894935 - Disable textarea resize based on popuppositioned  r=translations-reviewers,desktop-theme-reviewers,gregtatum,emilio
Determines if the SelectTranslationsPanel was positioned before
opening such that the bottom corner of the panel is the anchor
point. This causes the text-area resizer to grow the panel
upward even though the user is dragging downward, so we should
disable resizing in this case.

Differential Revision: https://phabricator.services.mozilla.com/D209396

UltraBlame original commit: aa76a8a2f1937a81c6312bde90e28055489e76ad