Bug 1897770 - Move ESLint no-insecure-url definitions to general-test.js. r=frontend-codestyle-reviewers,Gijs
This moves the general test definition to be first, so that the other test types can build on it.
It also re-arranges the no-insecure-url rule definitions to be defined in a single place in general-test.

This also exposes a few more places where we're not currently meeting the no-insecure-url rule, so we allowlist
those as well for now.

Differential Revision: https://phabricator.services.mozilla.com/D210927

UltraBlame original commit: 66cd1e7d29af184061ee0fd2cea7e6da2156c353