Bug 1895521 - fix fxa telemetry in the main browser UI. r=skhamis
browser-sync.js has got a bit disorganized over time. This patch tries to consolidate
all checks for which part of the UI is being interacted with and ensures both client
and server telemetry agree on the interaction, and as a part of this, ensures that
interactions with the fxa-toolbar item and the app menu are reported correctly.

Differential Revision: https://phabricator.services.mozilla.com/D209696

UltraBlame original commit: e76bee013d68292dbb3f1bbb03206fe09969f6c6