Bug 1897201 - Return identity transform while recovering from context loss with canvas. r=gfx-reviewers,lsalzman
When we lose the context for a 2D canvas, we need to silently fail any
canvas calls until the context is recovered. Unfortunately that meant we
returned a nullptr in CanvasRenderingContext2D::GetTransform which is
unsupported by the WebIDL. For now, let's just return an identity
transform as a placeholder for these situations.

Differential Revision: https://phabricator.services.mozilla.com/D211471

UltraBlame original commit: c82f94b991664b361e4d0abfc407e03d6a91212b