Bug 1899138 - Set the IsInThirdPartyContext for the loadInfo of the channel that loads worker scripts. r=asuth
To set the IsInThirdPartyContext bit to the channel's loadInfo, we
duduce whether the channel is foreign by looking into the partitionKey
of the worker. If the partitionKey of the OAs of the worker's principal
is not empty, the worker should be foreign, so as to the channel.

Differential Revision: https://phabricator.services.mozilla.com/D211757

UltraBlame original commit: 5354ecb8e5a2c2a88cfdd1304da79de45293f2df