Bug 1850834 - Teach the frame constructor to clean up anonymous wrappers. r=dholbert
So that removing stuff under an anonymous box (such like emptying a table that
has an anonymous table cell) doesn't have to re-create the table itself, which
could trigger pathological behavior if that needs to remove more stuff (due to
it being inside an ib-split for example).

To do that, unify how we deal with anonymous wrapper boxes a little bit, which
avoids the need to special-case the grid/flex anonymous items too, and
optimizes them in the same way. E.g., emptying a div like:

  <div style="display: flex">Some random text</div>

Will no longer reframe the <div>, just remove the text and the anonymous
wrapper.

We need to remove the eTypeBlock check added in bug 500467 since now removing a
table-related pseudo might in fact need white-space to be suppressed or
un-suppressed. This is tested by some of the table-anonymous-boxes reftests.

Differential Revision: https://phabricator.services.mozilla.com/D212292

UltraBlame original commit: 8f5e711b42ae6770d416aeeb93a76d74092f89a1