Bug 1900085 - Disable builtin Glean pings during tests r=TravisLong,perftest-reviewers,sparky,search-reviewers,jteow
This patch specifically disables "baseline" and "events" pings through not
telling Glean about user (in)activity, which triggers their submission.
"baseline" should matter less, but if an "events" ping is submitted during a
test it could contain a test's recorded events that it will later want to
assert are present.

"metrics" pings are already suppressed in automation via the `MOZ_AUTOMATION`
envvar. So update the docs to note that.

"deletion-request" pings are permitted at all times for all reasons and are not
included in the builtin pings we're disabling.

Differential Revision: https://phabricator.services.mozilla.com/D212315

UltraBlame original commit: f5c177cd9071dccc3d9a9f0bfe51962136f1cd46