Bug 1615282 - Stop screencast when there is non-recoverable error r=pehrsons
In DesktopCapture, the callback OnCaptureResult also returns a result of
CaptureFrame, whether it was successful or failed. When ERROR_PERMANENT
is returned, it means that all the following attempts to capture a frame
will not succeed and at this point we should stop capturing. To stop
capturing, we propagate CaptureEnded event up to the DeviceListener and
stop screencast.

Differential Revision: https://phabricator.services.mozilla.com/D201956

UltraBlame original commit: 12a0431a52e4fbe73143c647d45fc5d98ee7c029