Bug 1900127 - Clean up SelectTranslations default states r=translations-reviewers,ohall
Removes the default focus outline when encountering error
states in the SelectTranslationsPanel, such as on the
try-again button. UX recommendation is that the button
is focused, but not outlined, unless the user tabbed to it.
Also ensures that the textarea is always cleared before reopening.

Differential Revision: https://phabricator.services.mozilla.com/D212739

UltraBlame original commit: 9828d6d4e8f688fedc46625c7ebcfec09c0c466f